requirements.txt
bcrypt==4.0.1
cffi==1.15.1
colorama==0.4.5
cryptography==38.0.1
loguru==0.6.0
paho-mqtt==1.6.1
paramiko==2.11.0
pycparser==2.21
PyNaCl==1.5.0
six==1.16.0
win32-setctime==1.1.0
